balance updates after changing the expense

+++ HaushaltServices/src/main/java/com/arash/model/UserEntity.java
@@ -42,7 +42,7 @@ public class UserEntity {
 
 	// Getter and setter methods
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
